requirements.txt
asgiref==3.5.2
autopep8==2.0.0
black==22.10.0
boto3==1.26.24
botocore==1.29.24
cachetools==5.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
decorator==5.1.1
dj-database-url==1.0.0
Django==4.1.4
django-cors-headers==3.13.0
django-dotenv==1.4.2
django-filter==22.1
django-jazzmin==2.6.0
django-storages==1.13.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
future==0.18.2

google-auth==2.15.0
gunicorn==20.1.0
idna==3.4
jmespath==1.0.1
mypy-extensions==0.4.3
pathspec==0.10.2
Pillow==9.3.0
platformdirs==2.6.0
psycopg2==2.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.10.0
PyJWT==2.6.0
python-dateutil==2.8.2
pytz==2022.6
ratelim==0.1.6
requests==2.28.1
rsa==4.9
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
tzdata==2022.7
urllib3==1.26.13
whitenoise==6.2.0
